% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PitchersAllFactorsRatedCsvSaveFunction.R
\name{AllFactorsRatedModelPitchers}
\alias{AllFactorsRatedModelPitchers}
\title{Simple function obtains All Factors Rated modeling of Pitchers for desired date}
\usage{
AllFactorsRatedModelPitchers(Date = "8_9_2016")
}
\arguments{
\item{Date}{M_D_YYYY}
}
\value{
saved csv file of the model
}
\description{
Simple function obtains All Factors Rated modeling of Pitchers for desired date
}
\examples{
AllFactorsRatedModelPitchers(Date="8_9_2016")
}

