% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineNFLModelsFunction.R
\name{combineNflModels}
\alias{combineNflModels}
\title{Bind NFL Models}
\usage{
combineNflModels(x)
}
\arguments{
\item{x}{}
}
\value{
binded dataframe of all weeks in season
}
\description{
Bind NFL Models
}
\examples{
combineNflModels(allModels)
}

