% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFLCashLineDataFunction.R
\name{dfsGoldNflRecap}
\alias{dfsGoldNflRecap}
\title{NFL CashLine Contest Data}
\usage{
dfsGoldNflRecap(i, year = "2016")
}
\arguments{
\item{i:}{week for data}
}
\value{
Dk NFL cash line and winning scores; saves as .csv for later use as well
}
\description{
NFL CashLine Contest Data
}
\examples{
nflWeek2Cashlines <- dfsGoldNflRecap(2)
}

