% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract10FromModel.R
\name{extract10Mlb}
\alias{extract10Mlb}
\title{Create a new data frame from a Fantasy Lab MLB player model data frame.}
\usage{
extract10Mlb(n1, n2, n3, n4, n5, n6, n7, n8, n9, n10, model = "mlbModel")
}
\arguments{
\item{n1}{Name of a player from a player model data frame}

\item{n2}{Name of second player}

\item{n3}{Name of third player}

\item{n4}{Name of fourth player}

\item{n5}{Name of fifth player}

\item{n6}{Name of sixth player}

\item{n7}{Name of seventh player}

\item{n8}{Name of eighth player}

\item{n9}{Name of ninth player}

\item{n10}{Name of tenth player}

\item{model}{Data frame model to pull from}
}
\value{
Updated data frame containing only the 10 players in question.
}
\description{
Receives 10 player names to create a data frame with for further analysis.
}
\examples{
extract10Mlb(n1="David Ortiz",n2="Jose Altuve",n3="Neil Walker",n4="Mike Trout",n5="Bryce Harper",n6="Curtis Granderson",n7="Lucas Duda",n8="Alex Dickerson",n9="Trea Turner",n10="Freddie Freeman", model=mlbModel)
}

