% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashLineFunction.R
\name{getCashLine}
\alias{getCashLine}
\title{Function to lookup past cash lines(DK)}
\usage{
getCashLine()
}
\value{
all slate cash lines for given date
}
\description{
Function to lookup past cash lines(DK)
}
\examples{
getCashLine()
}

