% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashLineLoop.R
\name{getCashLineLoop}
\alias{getCashLineLoop}
\title{CashLine function for multiple iterations}
\usage{
getCashLineLoop(urlDate = "aug-12-2016")
}
\arguments{
\item{urlDate}{}
}
\value{
cashLines for all slates
}
\description{
CashLine function for multiple iterations
}
\examples{
getCashLineLoop(urlDate = "aug-i-2016")
}

