% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbPlayerModelCurlFunction.R
\name{getMlbPlayerModel}
\alias{getMlbPlayerModel}
\title{Fantasy Labs MLB model importing function}
\usage{
getMlbPlayerModel()
}
\value{
The model as a data frame
}
\description{
Import FantasyLabs MLB player model as a data frame; interactively input the date and a selection of hitters or pitchers.
}
\examples{
getMlbPlayerModel()
}

