% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitterModelForEachStructure.R
\name{getMlbPlayerModelLoop}
\alias{getMlbPlayerModelLoop}
\title{FANTASY labs model retrieve loop}
\usage{
getMlbPlayerModelLoop(modelDate = "8_1_2016", id = "301")
}
\arguments{
\item{modelDate}{as 4_5_2016}

\item{id}{301/302 hitter/pitcher}
}
\value{
FantasyLabs data frame
}
\description{
FANTASY labs model retrieve loop
}
\examples{
getMlbPlayerModelLoop(modelDate = "8_1_2016", id ="301")
}

