% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFLBasicLabsModelRetrieve.R
\name{getNflLabModel}
\alias{getNflLabModel}
\title{NFL FantasyLabs Model Retrieve}
\usage{
getNflLabModel(modelYear = "2015", modelWeek = NULL, model = NULL,
  mod = "", PositionID = NULL,
  cookie = "__cfduid=d19da73f21ae9852e137438d8ffd398371475796569; LD_U=http\%3A\%2F\%2Fwww.fantasylabs.com\%2F; .AspNet.Cookies=4OqHDcdGDsIxvmTJlDZP4l5WDNavz_EqYieZYNQ0jwz5D4hdcpnW81hhl10guRKYUJ4FUaeWfGILkrlHYAc5iAfkM78TCuId4C3Msk1UXmSOrCRGtWWf8LD2PvJVzmQDYuhLPJxChRIyLrl1J_wmMfFVG1GmepgwOUhW6njdFz7iNy7KAp-QjTnS7IQs20QabkcogIBaqqNQDyuNHD-7oeljqkv1ZDJq6CuGlLxe_8VmH8IX9kEtk8WM2GWATZaahPx09bCoYiNuWcydMTVG2vEOxBjy_obw7EgcJaQ3L8BjTRtfezcSo3_zRaSbSd8hlzFdC3gXQjm5x3H48o68HzJ6Moj3FjwIiAE7cv2KOoZ8GqJBgdD-aSfYangaMtwkgSlO4wYb2eOiLkjZojaBf-nVARxa89E293Fan6i4dZXyCTetik94qijmISHokbYG01lOQUoEPWjm07GsoqWvEp6AVplqN5DVHnqiWC23e54; flid=EriUJhQuj0yimw3ajUgOew; __distillery=a88774e_344d8f47-a8f7-4094-81ba-d3336d4c2d4d-ba08c6bef-19ba082fafba-d101; LD_R=http\%3A\%2F\%2Fwww.fantasylabs.com\%2Fnfl\%2Fplayer-models\%2F; __zlcmid=cyfuDnxjy8XwIK; LD_S=1475989257516; _gat=1; _ga=GA1.2.407536382.1475796573; LD_T=6a763f78-d230-4827-f05d-37ca01630a5d'")
}
\arguments{
\item{modelYear}{}

\item{modelWeek}{}

\item{model}{}

\item{mod}{}

\item{PositionID}{}
}
\value{
FantasyLabs NFL model
}
\description{
NFL FantasyLabs Model Retrieve
}
\examples{
getNflLabModel(modelYear = "2015",modelWeek = NULL,model = NULL,mod = "", PositionID = NULL)
}

