% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFLLinearCombosOmmissionFunction.R
\name{linearModelNFLSetup}
\alias{linearModelNFLSetup}
\title{NFL Linear Model Fitting Function}
\usage{
linearModelNFLSetup(NFLModel = DST2015SeasonAllBinded)
}
\arguments{
\item{NFLModel}{}
}
\value{
LM for NFL
}
\description{
NFL Linear Model Fitting Function
}
\examples{
...
}

