% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lModelbyTeamsFunction.R
\name{lModelTeams}
\alias{lModelTeams}
\title{labs Model Grouping}
\usage{
lModelTeams(lModel = lModel)
}
\arguments{
\item{lModel}{}
}
\value{
grouped model
}
\description{
labs Model Grouping
}
\examples{
lModelTeams(lModel=lModel)
}

