% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLB_Research_Full.R
\name{MLB_Data_Scripted_Function}
\alias{MLB_Data_Scripted_Function}
\title{Full Run of MLB Tools}
\usage{
MLB_Data_Scripted_Function(Date = "7_20_2016")
}
\arguments{
\item{Date}{as M_d_YYYY}
}
\value{
several csv files saved to computer
}
\description{
Full Run of MLB Tools
}
\examples{
MLB_Data_Scripted_Function(Date = "7_20_2016")
}

