% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLBFullLoop.R
\name{MLB_Data_Scripted_Function_Loop}
\alias{MLB_Data_Scripted_Function_Loop}
\title{Full running MLB tool on a loop}
\usage{
MLB_Data_Scripted_Function_Loop(Ddate = paste0("6_", i, "_2016"),
  modelDateFlabs = "6_12_2016", idFlabs = "301", gDateSwisher = "06",
  urlDateCashLine = "jun")
}
\arguments{
\item{Ddate}{month and year with day being looped}

\item{modelDateFlabs}{}

\item{idFlabs}{301 for hitters 302 for pitchers}

\item{gDateSwisher}{month mm format}

\item{urlDateCashLine}{month (3 letter format)}
}
\value{
model with swish ratings(hitters)
}
\description{
Full running MLB tool on a loop
}
\examples{
foreach(i=1:30) \%do\% MLB_Data_Scripted_Function_Loop(Ddate=paste0("6_",i,"_2016"), modelDateFlabs = "6_12_2016",idFlabs ="301", gDateSwisher = "06", urlDateCashLine = "jun")
}

