% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlatedModelDataFunction.R
\name{mlbCorrToCsv}
\alias{mlbCorrToCsv}
\title{Export Fantasy Labs player model correlated data as csv}
\usage{
mlbCorrToCsv(model = mlbModel, y = "MLB_date_correlated")
}
\arguments{
\item{model}{Name of player model}

\item{y}{Name to save the file as}
}
\value{
The name of the csv filed saved to your computer.
}
\description{
Calculates the correlation between all objects and variables in a player model and saves it as a csv to your working directory.
}
\examples{
mlbCorrToCsv(model=mlbModel,y="MLB_date_correlated")
}

