% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlynumericfunction.R
\name{mlbNumericOnly}
\alias{mlbNumericOnly}
\title{Prepare Fantasy Labs MLB model for more analysis}
\usage{
mlbNumericOnly(model = mlbModel)
}
\arguments{
\item{model}{Name of player model}
}
\value{
none
}
\description{
Takes a player model data frame and returns only the numeric data for analysis with librarys such as corrplot
}
\examples{
mlbNumericOnly(model=mlbModel)
}

