% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflCorrelationsFunction.R
\name{NFLCorrelations}
\alias{NFLCorrelations}
\title{NFL Correlations}
\usage{
NFLCorrelations(x)
}
\arguments{
\item{x}{}
}
\value{
correlation DF
}
\description{
NFL Correlations
}
\examples{
....
}

