% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFLcsvModelImportFunction.R
\name{nflModelImportfromCSV}
\alias{nflModelImportfromCSV}
\title{NFL CSV Model Import}
\usage{
nflModelImportfromCSV(modelYear = "2015", modelWeek = "NULL",
  PositionID = "QB", mod = NULL)
}
\arguments{
\item{modelYear}{}

\item{modelWeek}{}

\item{PositionID}{}

\item{mod}{}
}
\value{
converted csv model to dataframe in current r session
}
\description{
NFL CSV Model Import
}
\examples{
nflModelImportfromCSV(modelYear = "2015",modelWeek = "NULL",PositionID = "QB",mod = NULL)
}

