% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffWR_CB_function.R
\name{pffWR_CB}
\alias{pffWR_CB}
\title{ProFootballFocus WR/CB Chart Function}
\usage{
pffWR_CB(week)
}
\arguments{
\item{week}{}
}
\value{
WR/CB Chart
}
\description{
ProFootballFocus WR/CB Chart Function
}
\examples{
WR_CBWk1 <- pffWR_CB(1)
}

