% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPlayerModelFunction.R
\name{playerModelFilter}
\alias{playerModelFilter}
\title{A filtering option for FantasyLab MLB models}
\usage{
playerModelFilter(name, model = h.7_23, filter = "actual")
}
\arguments{
\item{Name}{Player's name}

\item{Model}{Name of data frame}

\item{Filter}{name of stat to find}
}
\value{
The player stat desired according to your model
}
\description{
Get specific stat for one player in a fantasylabs player model data frame.
}
\examples{
playerModelFilter("Yoenis Cespedes",model=mlbModel,filter="actual")
}

