% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeColFunction.R
\name{removeCol}
\alias{removeCol}
\title{Remove DataFrame Column}
\usage{
removeCol(df, col)
}
\arguments{
\item{df}{}

\item{col}{}
}
\value{
Data Frame lacking desired column
}
\description{
Remove DataFrame Column
}
\examples{
df <- removeCol(df,col)
}

