% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sModelbyTeamRatingsFunction.R
\name{sModelByTeamRatings}
\alias{sModelByTeamRatings}
\title{SwisherModel groups}
\usage{
sModelByTeamRatings(sModel = sModel)
}
\arguments{
\item{sModel}{}
}
\value{
model grouped
}
\description{
SwisherModel groups
}
\examples{
sModelByTeamRatings(sModel = sModel)
}

