% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swisherly.R
\name{swisherlyLabModel}
\alias{swisherlyLabModel}
\title{The full flow}
\usage{
swisherlyLabModel(swishModel = sModel, labsModel = lModel)
}
\arguments{
\item{swishModel}{}

\item{labsModel}{}
}
\value{
all
}
\description{
The full flow
}
\examples{
swisherlyLabModel(SwishModel = sModel, labsModel = lModel)
}

