% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swisherlyLabModelLoop.R
\name{swisherlyLabModelLoop}
\alias{swisherlyLabModelLoop}
\title{Looping function for the swish batter ratings}
\usage{
swisherlyLabModelLoop(swishModel = sModel, labsModel = lModel)
}
\arguments{
\item{swishModel}{}

\item{labsModel}{}
}
\value{
a set of projections
}
\description{
Looping function for the swish batter ratings
}
\examples{
swisherlyLabModelLoop(swishModel = sModel, labsModel = lModel)
}

