% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSwishHitters.R
\name{swishHitters}
\alias{swishHitters}
\title{Retreive the batter comparison table from Swish Analytics}
\usage{
swishHitters()
}
\value{
swish data frame
}
\description{
Runs as an interactive function asking for a Date and returns the Swish table as a data frame.
}
\examples{
swishHitters()
}

