% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swishHittersLoop.R
\name{swishHittersLoop}
\alias{swishHittersLoop}
\title{Simple Swish rating for a loop}
\usage{
swishHittersLoop(gDate = "2016-08-12")
}
\arguments{
\item{gDate}{as yyyy-mm-dd}
}
\value{
a simplified swish rating for hitters
}
\description{
Simple Swish rating for a loop
}
\examples{
swishHitterLoop(gDate = "2016-08-12")
}

