% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFLVIFLinearModelFitting.R
\name{vifStepLinearModelNFL}
\alias{vifStepLinearModelNFL}
\title{NFL VIF fitting}
\usage{
vifStepLinearModelNFL(QBFit, yQB, xQB)
}
\arguments{
\item{QBFit}{}
}
\value{
VIFS Linear models
}
\description{
NFL VIF fitting
}
\examples{
...
}

