% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelRetrieveWindowsFunction.R
\name{windowsNFLModel}
\alias{windowsNFLModel}
\title{NFL model for Windows}
\usage{
windowsNFLModel(userName = "Benjaminryanshopping@icloud.com",
  passWord = "W3ytjk589682", modelYear = "2016", modelWeek = 1,
  modelPosition = "QB")
}
\arguments{
\item{userName}{}

\item{passWord}{}

\item{modelYear}{}

\item{modelWeek}{}

\item{modelPosition}{}
}
\value{
NFLplayerModel
}
\description{
NFL model for Windows
}
\examples{
NFLQB2016Wk1 <- windowsNFLModel(userName = "Benjaminryanshopping@icloud.com",passWord = "W3ytjk589682",modelYear = "2016",modelWeek = 1,modelPosition = "QB")
}

