% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFLBasicLabsModelRetrieve.R
\name{getNflLabModel}
\alias{getNflLabModel}
\title{NFL FantasyLabs Model Retrieve}
\usage{
getNflLabModel(modelYear = "2015", modelWeek = NULL, model = NULL,
  mod = "", PositionID = NULL)
}
\arguments{
\item{modelYear}{}

\item{modelWeek}{}

\item{model}{}

\item{mod}{}

\item{PositionID}{}
}
\value{
FantasyLabs NFL model
}
\description{
NFL FantasyLabs Model Retrieve
}
\examples{
getNflLabModel(modelYear = "2015",modelWeek = NULL,model = NULL,mod = "", PositionID = NULL)
}

